React Maintenance: Performance | Component files moved into /components branch and css file moved into /css branch
